pyproject.toml
[tool.poetry]
name = "VolPy"
version = "0.1.0"
description = ""
authors = ["Clément POIRET <poiret.clement@outlook.fr>"]

[tool.poetry.dependencies]
python = ">3.8"
hydra-core = "^1.1.1"
beautifulsoup4 = "^4.11.1"
requests = "^2.27.0"
lxml = "^4.8.0"

[tool.poetry.scripts]
volpy = "volpy.vol:main"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
